test: merge local test Configuration into test
